% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{cor.mtest}
\alias{cor.mtest}
\title{A correlation statistical test taking a matrix}
\usage{
cor.mtest(mat, ...)
}
\arguments{
\item{mat}{: a matrix of quantitatives values}

\item{...}{: Any usual parameter for the cor.test function from R}
}
\value{
return the coefficient of correlation
}
\description{
This function will compute the pvalues of a correlation test between variables in a matrix
}
\details{
Call example : correlation(data, "group", "X.tissus")
}
