% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{correlation}
\alias{correlation}
\title{Global correlation method}
\usage{
correlation(dataframe, var1, var2)
}
\arguments{
\item{dataframe}{: a dataframe from list_type function.}

\item{var1}{: The name of a variable, between double quotes}

\item{var2}{: The name of a variable, between double quotes}
}
\value{
return the coefficient of correlation
}
\description{
This function will compute the coefficient of correlation of two variables (given by their names)
and according to their type.
This function will be automatically called by the modelise function if it detects that there is only one explanation variable.
}
\details{
Call example : correlation(data, "group", "X.tissus")
}
