% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{create_correlation_table}
\alias{create_correlation_table}
\title{A correlation table function}
\usage{
create_correlation_table(data, cutoff, level = 0.95, na, method,
  computepvalues)
}
\arguments{
\item{data}{: a dataframe with variables in column and numeric values}

\item{cutoff}{: a value of absolute difference or similarity that will choose if edges are included in the edgelist or not}

\item{level:}{the percentage of confidence (0.95 for 95%)}

\item{na:}{default is "omit", select the na action}

\item{method:}{default is "pearson", the name of the correlation technique. available: pearson, spearman, kendall, bicor, hoeffding}

\item{computepvalues:}{default is FALSE, if you want the function to compute pvalues set TRUE. Computation will be longer.}
}
\value{
return a edge list table
}
\description{
This function will compute the coefficients of correlation and the associated p-values between all variables in the dataframe
}
\details{
Call example : create_correlation_table(dataframe)
}
