% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{create_correlation_table_double_df}
\alias{create_correlation_table_double_df}
\title{A correlation table function}
\usage{
create_correlation_table_double_df(data, data2 = NULL, cutoff, level = 0.95,
  na, method = "pearson", computepvalues = FALSE, type = "differential")
}
\arguments{
\item{data}{: a dataframe with variables in column and numeric values}

\item{data2}{: a second dataframe with variables in column and numeric values}

\item{cutoff}{: a value of absolute difference or similarity that will choose if edges are included in the edgelist or not}

\item{type}{: default is "differential", in differential mode, edges with a abs(coef1-coef2) > cutoff will be included in the edge list. If type = "common", edges with abs(coef1-coef2) < cutoff will be included.}

\item{level:}{the percentage of confidence (0.95 for 95%)}

\item{na:}{default is "omit", select the na action}

\item{method:}{default is "pearson", the name of the correlation technique. available: pearson, spearman, kendall, bicor, hoeffding}

\item{computepvalues:}{default is FALSE, if you want the function to compute pvalues set TRUE. Computation will be longer.}
}
\value{
return a edge list table
}
\description{
This function will compute the coefficients of correlation and return a differential or common edge list
}
\details{
Call example : create_correlation_table_double_df(groupA, groupB, 0.7, 0.95, "omit", "spearman", FALSE, "common")
}
