% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_report}
\alias{create_report}
\title{Report creation for IPToolbox R package}
\usage{
create_report(dataframe, dependantvar, explanationvar, directory,
  reportname = NULL)
}
\arguments{
\item{dataframe}{: The dataframe used in the model the user wants to report}

\item{dependantvar}{: The name of the dependant variable used in the model the user wants to report.}

\item{explanationvar}{: The list of explanation variable used in the model the user wants to report.}

\item{directory}{: The working directory (where the Rdata file is located) (Currently seeking an alternative to this method)}
}
\value{
An invitation to comile the .tex file. The autocompilation in a pdf is implemented but if the user doesn't have pdftex it will fail.
}
\description{

}
\details{
Called by the model function
}
