% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hub_bottlenecks}
\alias{hub_bottlenecks}
\title{HUB and Bottlenecks computation function}
\usage{
hub_bottlenecks(graph, filename.hub, filename.bottlenecks, top = TRUE)
}
\arguments{
\item{graph}{: A igraph object}

\item{filename.hub}{: the name of hubs files whithout extension}

\item{filename.bottlenecks}{: the name of bottlenecks files whitout extention}

\item{top}{: default = true, if top is true compute the top 100 of hubs and bottlenecks (if there is more than 100 hubs and/or 100 bottlenecks)}
}
\value{
write tsv files with hubs and bottlenecks
}
\description{

}
\details{

}
