% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{learning_sample}
\alias{learning_sample}
\title{Random sampling in a dataframe in order to create learning and validation samples}
\usage{
learning_sample(dataframe, reset)
}
\arguments{
\item{dataframe}{: a dataframe with raw data : individuals are in rows, variables in columns}

\item{reset}{: TRUE if we want to replace the drawn value, FALSE otherwise}
}
\value{
return the learning sample in a dataframe
}
\description{

}
\details{

}
