% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_type}
\alias{list_type}
\title{List the type of variables in a dataframe}
\usage{
list_type(dataframe)
}
\arguments{
\item{dataframe}{a dataframe with raw data}
}
\value{
a object containing a list and the input dataframe
}
\description{
This function analyse variables types avalaible in the dataframe
These types are added to a list
The order of this list is the order of variables in the dataframe
}
