% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.R
\name{model.linear}
\alias{model.linear}
\title{linear modelisation method}
\usage{
model.linear(data.model, data.predict, dependantvar, level = 0.95,
  explanationvar, switch)
}
\arguments{
\item{data.predict}{: a dataframe  with data to predict from learning.sample}

\item{dependantvar}{: the name between quotes of the dependant variable : values needs to be continuous}

\item{level}{: the level of confidence for the prediction ex: 0.95 (equal alpha risk 5 percent).}

\item{explanationvar}{: a vector with the names of columns refering to explanation variables.}

\item{switch}{: If TRUE and dependantvar}

\item{dataframe}{: a dataframe from list_type function.}
}
\value{
return the model, objects with data for the report creation.

if the dependant variable is binary, then also return roc object
}
\description{

}
\details{
Call example : model.linear(data, data.predict, "group", 0.95, vecteur_explanationvar, FALSE).
}
