% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.R
\name{model.multinomial.logistic}
\alias{model.multinomial.logistic}
\title{Multinomial Logistic regression function}
\usage{
model.multinomial.logistic(data.model, data.predict, dependantvar,
  level = 0.95, explanationvar, switch, response)
}
\arguments{
\item{dependantvar}{: the name between quotes of the dependant variable : value needs to be binary or categorial}

\item{level}{: the level of confidence for the prediction ex: 0.95 (equal alpha risk 5 percent).}

\item{explanationvar}{: a vector with the names of columns refering to explanation variables.}

\item{dataframe}{: a dataframe from list_type function.}

\item{mode}{: the sample technique to use.}
}
\value{
return the model, objects with data for the report creation.

if the dependant variable is binary, then also return roc object
}
\description{

}
\details{
Call example : model.logistic(data, "group", 0.95, vecteur_explanationvar).
}
