% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{modelise}
\alias{modelise}
\title{Global modelisation method}
\usage{
modelise(dataframe, dependantvar, colexplanation, datatopredict, level)
}
\arguments{
\item{dataframe}{: a dataframe from list_type function}

\item{dependantvar}{: the name between quotes of the dependant variable}

\item{colexplanation}{: numbers of columns corresponding to the explanation variables to include in the model}

\item{datatopredict}{: a datafram with only the data to predict}

\item{level}{: the level of confidence for the prediction ex: 0.95 (equal alpha risk 5%)}
}
\value{
return the predicted results and print AUC value or Model Accuracy percentage
}
\description{

}
\details{
Call exemple : modelise (data, "group", c(1,8:16), data.predict, 0.95)
}
