% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{modelise2}
\alias{modelise2}
\title{Global modelisation method 2}
\usage{
modelise2(dataframe, dependantvar, mode = "everyone", level = 0.95,
  explanationvar)
}
\arguments{
\item{dataframe}{: a dataframe from list_type function.}

\item{dependantvar}{: the name between quotes of the dependant variable.}

\item{mode}{: the sample technique to use.}

\item{level}{: the level of confidence for the prediction ex: 0.95 (equal alpha risk 5%).}

\item{explanationvar}{: a vector with the names of columns refering to explanation variables.}
}
\value{
return the predicted results and print AUC value or Model Accuracy percentage
}
\description{

}
\details{
Call example : modelise2(data, "group", "random_test", 0.95, vecteur_explanationvar).
mode take "everyone" by default. everyone mean that the predict will be made on every available patient.
if mode = random_reset, a learning sample with reset will me made and the predict data will be the validation sample.
if mode = random, a learning sample will me made by randomly sampling patients in the dataframe. Predict data will be the validation sample.
if mode = a vector with patients id's for the learning sample, these patients will be in the learning sample and the predict data will be the validation sample.
}
