% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Number.ROC}
\alias{Number.ROC}
\title{Number of ROC curves for a repsonse vector}
\usage{
Number.ROC(response)
}
\arguments{
\item{response:}{a vector containg each possible response}
}
\value{
return the number of ROC curves (possible unique association in response)
}
\description{

}
\details{

}
