% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{Pearson}
\alias{Pearson}
\title{Pearson's Coefficient of correlation}
\usage{
Pearson(X, Y, na.rm = TRUE)
}
\arguments{
\item{X}{: A quantitative variable}

\item{Y}{: Another quantitative variable}
}
\value{
The pearson coefficient of correlation value
}
\description{

}
\details{

}
