% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.coexpression}
\alias{plot.coexpression}
\title{A co-expression network plotting function}
\usage{
plot.coexpression(edge.table, directed = TRUE, title,
  matrix.foldchange = NULL, color = "fold-change", directory,
  filename = "random_filename.pdf", vertex.size = 3.5, vertex.label = NA,
  edge.arrow.size = NA)
}
\arguments{
\item{edge.table}{: a edge table (or edge list) from the function create_correlation_table (or create_correlation_table_double_df)}

\item{directed}{: Takes TRUE (default) if the graph as to be directed or FALSE if not}

\item{title}{: the title of the graph}

\item{matrix.foldchange}{: default is NULL, is coloring technique is fold-change then a matrix of fold-change is needed}

\item{color}{: default is "fold-change", define the color technique}

\item{directory}{: set a path where to store the pdf}

\item{filename}{: default is "random_filename.pdf", set the name of the pdf file}

\item{vertex.size}{: default is 3.5, set the size of nodes in the co-expression network}

\item{vertex.label}{: default is NA, set the labels of nodes in the co-expression network}

\item{edge.arrow.size}{: default is NA, set the arrow size in the co-expression network}
}
\value{
return a dataframe
}
\description{
This function will plot a co-expression network
}
\details{
Call example : plot.coexpression(myedgetable, "This is the title of my co-expression network", mymatrixoffoldchange, color = "fold-change", "path to a folder", "The name of the file")
}
