% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.correlogram}
\alias{plot.correlogram}
\title{A correlogram plotting function}
\usage{
plot.correlogram(pdf.name, matrix.coef, matrix.pvalues, level = 0.95,
  method = "number", number.size = 0.35, pvalues.size = 0.3,
  title = pdf.name, mar = c(0, 0, 1, 0), label.size = 0.5)
}
\arguments{
\item{pdf.name}{: the name of the pdf file (and by default also title of the plot)}

\item{matrix.coef}{: A square matrix with all coefficients from square_matrix_from_edge_list}

\item{matrix.pvalues}{: A square matrix with all pvalues from square_matrix_from_edge_list}

\item{level}{: the value of confidence (0.95 by default)}

\item{method}{: the corrplot method, number by default}

\item{number.size}{: a numeric value that will act as a multiplier on the number size}

\item{pvalues.size}{: a numeric value that will act as a multiplier on the pvalues size}

\item{title}{: the title of the plot, by default the pdf anme}

\item{mar}{: margin values. By default 0,0,1,0}

\item{label.size}{: a numeric value that will act as a multiplier on the label size}
}
\value{
return a corrplot
}
\description{
This function will call the correlogram plotting function from corrplot package
}
\details{
Call example : plot.correlogram("mypdfname",matric.coefficients, matrix.pvalues, 0.95, "number")
}
