% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{Spearman}
\alias{Spearman}
\title{Spearman's Coefficient of correlation}
\usage{
Spearman(X, Y, na.rm = TRUE)
}
\arguments{
\item{X}{: continuous or ordinal variable}

\item{Y}{: continuous or ordinal variable}
}
\value{
The spearman coefficient of correlation value
}
\description{

}
\details{

}
