% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{square_matrix_from_edge_list}
\alias{square_matrix_from_edge_list}
\title{A square matrix creation function}
\usage{
square_matrix_from_edge_list(data, content = "coefficients")
}
\arguments{
\item{data}{: a edge list dataframe from create_correlation table function}

\item{content}{: default is "coefficients", determine what is in the square matrix}
}
\value{
return a square matrix
}
\description{
This function will compute a square matrix from an edgelist (from correlation tables functions)
}
\details{
Call example : square_matrix_from_edge_list(edge.list, "coefficients")
}
