% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validation_sample}
\alias{validation_sample}
\title{Create the validation sample from data not in the learning sample}
\usage{
validation_sample(learning_sample, dataframe)
}
\arguments{
\item{dataframe}{: The dataframe used with the Learning_Sample function}

\item{Learning_sample}{: The Learning dataframe creted with the Learning_Sample function}
}
\value{
return the validation dataframe associated with the parameter:  learning_sample
}
\description{

}
\details{

}
